Pulls PositionSolution & code out to own header file
